+++ RELEASE-NOTES.md
@@ -2,6 +2,10 @@
 
 **[Development Version](https://github.com/psidnell/ofexport2/archive/master.zip)**
 
+- Fixed bug in CSV template that marked all items as flagged.
+- Fixed a ranking bug that caused items to appear in the wrong order in Context mode or after flattening.
+- Improved Markdown formatting.
+
 **[1.0.11 (2014-12-10)](https://github.com/psidnell/ofexport2/archive/ofexport-v2-1.0.11.zip)**
 
 - Documentation Improvements